Migrate to .hpp
