+++ fuse_constraints/test/test_marginalize_variables.cpp
@@ -42,7 +42,7 @@
 #include <fuse_core/serialization.hpp>
 #include <fuse_core/uuid.hpp>
 #include <fuse_core/variable.hpp>
-#include <fuse_graphs/hash_graph.h>
+#include <fuse_graphs/hash_graph.hpp>
 #include <fuse_variables/orientation_3d_stamped.hpp>
 
 #include <boost/serialization/access.hpp>
