+++ fuse_publishers/test/test_pose_2d_publisher.cpp
@@ -35,7 +35,7 @@
 #include <fuse_core/eigen.hpp>
 #include <fuse_core/transaction.hpp>
 #include <fuse_core/uuid.hpp>
-#include <fuse_graphs/hash_graph.h>
+#include <fuse_graphs/hash_graph.hpp>
 #include <fuse_publishers/pose_2d_publisher.h>
 #include <fuse_variables/orientation_2d_stamped.hpp>
 #include <fuse_variables/position_2d_stamped.hpp>
