+++ fuse_publishers/test/test_stamped_variable_synchronizer.cpp
@@ -34,7 +34,7 @@
 #include <fuse_publishers/stamped_variable_synchronizer.h>
 
 #include <fuse_core/uuid.hpp>
-#include <fuse_graphs/hash_graph.h>
+#include <fuse_graphs/hash_graph.hpp>
 #include <fuse_variables/orientation_2d_stamped.hpp>
 #include <fuse_variables/position_2d_stamped.hpp>
